+++ uobtheatre/productions/schema.py
@@ -20,12 +20,11 @@ from uobtheatre.productions.models import (
     Warning,
     append_production_qs,
 )
-from uobtheatre.utils.schema import EnumNode, FilterSet, GrapheneEnumMixin
+from uobtheatre.utils.filters import FilterSet
+from uobtheatre.utils.schema import DjangoObjectType, GrapheneEnumMixin
 
 
 class CrewRoleNode(GrapheneEnumMixin, DjangoObjectType):
-    department = graphene.Field(EnumNode)
-
     class Meta:
         model = CrewRole
         interfaces = (relay.Node,)
